pyproject.toml
[tool.poetry]
name = "cohere-hackathon"
version = "0.1.0"
description = "tbd"
authors = ["Landon Edwards <jlandonedwards@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "1.26.1"
pandas = "*"
weaviate-client = "*"
cohere = "*"
matplotlib = "*"
seaborn = "*"
plotly = "*"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"